Error out when packet is empty
Error out when a explicitly mentioned packet is empty.

Any packet listed in the source package file now needs to exist. It is
still fine for their facets to stay empty, but the main packet must
exist!

It is still fine for an implicitly created packet that matches the
source package name to be empty though.