iGet rid of justconfig dependency
It is so unwieldy to use. Serde makes for slightly uglier config files,
but a lot less code.

The big feature of stacked configuration files is really cool, but not
what I expect users to want in this use case anyway. I expect them to
just write a new config file and not needing to do lots of custom
changes to distribution-provided files.